Case 273485 - Update dependencies
This gem doesn't specify specific Ruby versions but it fails to run on the 3.2
series becaues of old assumptions about option hashes vs keyword args. To
resolve I removed some overly tight restrictions on the dependencies and updated
all dependencies.

With that in place I could then make a few tweaks to tests such that we are
making clear to rspec that we are looking for option hashes instead of keyword
args.