Case 273485 - Improve callbacks to avoid NULL Auth0 ID
MO suggested that a subtle interaction with the user model might lead to
the Auth0 id being blanked out because the retrieval of the info from Auth0
and the saving of it locally is happening via communication of instance
variables and within an after_commit hook.

I don't think this is what happened but I don't know what happened and the
current setup here is unecessary. The "after_save" hook itself can just save
the attributes simplifying the code.

Therefore making that adjustments. I tried to update the tests as best I can.
They are heavily mocked (which I'm not a fan of) so I tried to reduce that a
bit by treating the behavior as a test on the caller not by calling the private
method directly. It's still fairly mocked so it doesn't give me a lot of
confidence. Therefore I also tested this manually against BloodHub by clearing
out the Auth0 ID (and picture). Then do a login and confirm they are
re-populated.